API reference

Basic structs

FastIsostasy.ComputationDomain — Type

ComputationDomain

Return a struct containing all information related to geometry of the domain and potentially used parallelism. To initialize one with 2*W and 2^n grid cells:

Omega = ComputationDomain(W, n)

FastIsostasy.PhysicalConstants — Type

PhysicalConstants

Return a struct containing important physical constants.

FastIsostasy.LateralVariability — Type

LateralVariability

Return a struct containing all information related to the radially layered structure of the solid Earth and its parameters.

Missing docstring.

Missing docstring for RefGeoState. Check Documenter's build log for details.

FastIsostasy.GeoState — Type

GeoState

Return a mutable struct containing the geostate which will be updated over the simulation.

FastIsostasy.PrecomputedFastiso — Type

PrecomputedFastiso(Omega::ComputationDomain, c::PhysicalConstants, p::LateralVariability)

Return a struct containing pre-computed tools to perform forward-stepping of the model, namely:

elasticgreen::AbstractMatrix{T}

fourier_elasticgreen::AbstractMatrix{T}{Complex{T}}

pfft::AbstractFFTs.Plan

pifft::AbstractFFTs.ScaledPlan

Dx::AbstractMatrix{T}

Dy::AbstractMatrix{T}

Dxx::AbstractMatrix{T}

Dyy::AbstractMatrix{T}

Dxy::AbstractMatrix{T}

negligible_gradD::Bool

rhog::T

geoidgreen::AbstractMatrix{T}

FastIsostasy.SuperStruct — Type

SuperStruct()

Return a struct containing all the other structs needed for the forward integration of the model:

Omega::ComputationDomain{T}

c::PhysicalConstants{T}

p::LateralVariability{T}

tools::PrecomputedFastiso{T}

Hice::Interpolations.Extrapolation

Hice_cpu::Interpolations.Extrapolation

eta::Interpolations.Extrapolation

eta_cpu::Interpolations.Extrapolation

refgeostate::RefGeoState{T}

geostate::GeoState{T}

interactive_geostate::Bool

Missing docstring.

Missing docstring for FastisoResults. Check Documenter's build log for details.

Mechanics

FastIsostasy.fastisostasy — Function

fastisostasy()

Main function. List of all available solvers here.

Missing docstring.

Missing docstring for forward_isostasy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for init_results. Check Documenter's build log for details.

Missing docstring.

Missing docstring for forwardstep_isostasy!. Check Documenter's build log for details.

FastIsostasy.dudt_isostasy! — Function

dudt_isostasy!()

Update the displacement rate dudt of the viscous response.

Missing docstring.

Missing docstring for simple_euler!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for corner_bc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for edge_bc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for compute_elastic_response. Check Documenter's build log for details.

Sea-level

FastIsostasy.update_geoid! — Function

update_geoid!(sstruct::SuperStruct)

Update the geoid by convoluting the Green's function with the load anom.

FastIsostasy.columnanom_ice — Function

columnanom_ice(sstruct)

Compute the density-scaled anomaly of the ice column w.r.t. the reference state.

FastIsostasy.columnanom_water — Function

columnanom_water(sstruct)

Compute the density-scaled anomaly of the (liquid) water column w.r.t. the reference state.

FastIsostasy.columnanom_mantle — Function

columnanom_mantle(sstruct)

Compute the density-scaled anomaly of the mantle column w.r.t. the reference state.

FastIsostasy.columnanom_load — Function

columnanom_load(sstruct)

Compute the density-scaled anomaly of the load (ice + liquid water) column w.r.t. the reference state.

FastIsostasy.columnanom_full — Function

columnanom_full(sstruct)

Compute the density-scaled anomaly of the all the columns (ice + liquid water + mantle) w.r.t. the reference state.

Correction of the surface distortion is not needed here since rho * A * z / A = rho * z.

FastIsostasy.get_geoidgreen — Function

get_geoidgreen(sstruct::SuperStruct)

Return the Green's function used to compute the anoms in geoid.

Reference

Coulon et al. 2021.

FastIsostasy.update_loadcolumns! — Function

update_loadcolumns!(sstruct::SuperStruct, u::AbstractMatrix{T}, H_ice::AbstractMatrix{T})

Update the load columns of a ::GeoState.

FastIsostasy.update_sealevel! — Function

update_sealevel!(sstruct::SuperStruct)

Update the sea-level by adding the various contributions.

Reference

Coulon et al. (2021), Figure 1.

FastIsostasy.update_slc! — Function

update_slc!(sstruct::SuperStruct)

Update the sea-level contribution of melting above floatation, density correction and potential ocean volume.

Reference

Goelzer et al. (2020), eq. (12).

FastIsostasy.update_V_af! — Function

update_V_af!(sstruct::SuperStruct)

Update the ice volume above floatation.

Reference

Goelzer et al. (2020), eq. (13). Note: we do not use eq. (1) as it is only a special case of eq. (13) that does not allow a correct representation of external sea-level forcings.

FastIsostasy.update_slc_af! — Function

update_slc_af!(sstruct::SuperStruct)

Update the sea-level contribution of ice above floatation.

Reference

Goelzer et al. (2020), eq. (2).

FastIsostasy.update_V_pov! — Function

update_V_pov!(sstruct::SuperStruct)

Update the potential ocean volume.

Reference

Goelzer et al. (2020), eq. (14). Note: we do not use eq. (8) as it is only a special case of eq. (14) that does not allow a correct representation of external sea-level forcinsstruct.geostate.

FastIsostasy.update_slc_pov! — Function

update_slc_pov!(sstruct::SuperStruct)

Update the sea-level contribution associated with the potential ocean volume.

Reference

Goelzer et al. (2020), eq. (9).

FastIsostasy.update_V_den! — Function

update_V_den!(sstruct::SuperStruct)

Update the ocean volume associated with the density correction.

Reference

Goelzer et al. (2020), eq. (10).

FastIsostasy.update_slc_den! — Function

update_slc_den!(sstruct::SuperStruct)

Update the sea-level contribution associated with the density correction.

Reference

Goelzer et al. (2020), eq. (11).

Parameter inversion

Missing docstring.

Missing docstring for ParamInversion. Check Documenter's build log for details.

FastIsostasy.perform — Function

perform(paraminv::ParamInversion)

Return priors and ukiobj that allow to extract the results of the parameter inversion as initialized in paraminv.

Missing docstring.

Missing docstring for where_response. Check Documenter's build log for details.

Utils

FastIsostasy.years2seconds — Function

years2seconds(t::Real)

Convert input time t from years to seconds.

FastIsostasy.seconds2years — Function

seconds2years(t::Real)

Convert input time t from seconds to years.

FastIsostasy.m_per_sec2mm_per_yr — Function

m_per_sec2mm_per_yr(dudt::Real)

Convert displacement rate dudt from $ m \, s^{-1} $ to $ mm \, \mathrm{yr}^{-1} $.

FastIsostasy.matrify — Function

matrify(x, Nx, Ny)

Generate a vector of constant matrices from a vector of constants.

FastIsostasy.get_r — Function

get_r(x::T, y::T) where {T<:Real}

Get euclidean distance of point (x, y) to origin.

FastIsostasy.meshgrid — Function

meshgrid(x, y)

Return a 2D meshgrid spanned by x, y.

FastIsostasy.dist2angulardist — Function

dist2angulardist(r::Real)

Convert Euclidean to angular distance along great circle.

Missing docstring.

Missing docstring for sphericaldistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scalefactor. Check Documenter's build log for details.

FastIsostasy.latlon2stereo — Function

latlon2stereo(lat, lon, lat0, lon0)

Compute stereographic projection (x,y) for a given latitude lat longitude lon, reference latitude lat0 and reference longitude lon0. Optionally one can provide lat::AbstractMatrix and lon::AbstractMatrix if the projection is to be computed for the whole domain. Note: angles must be provided in degrees! Reference: John P. Snyder (1987), p. 157, eq. (21-2), (21-3), (21-4).

FastIsostasy.stereo2latlon — Function

stereo2latlon(x, y, lat0, lon0)

Compute the inverse stereographic projection (lat, lon) based on Cartesian coordinates (x,y) and for a given reference latitude lat0 and reference longitude lon0. Optionally one can provide x::AbstractMatrix and y::AbstractMatrix if the projection is to be computed for the whole domain. Note: angles must be para elloprovided in degrees!

Convert stereographic (x,y)-coordinates to latitude-longitude. Reference: John P. Snyder (1987), p. 159, eq. (20-14), (20-15), (20-18), (21-15).

FastIsostasy.get_rigidity — Function

get_rigidity(t::T, E::T, nu::T) where {T<:AbstractFloat}

Compute rigidity D based on thickness t, Young modulus E and Poisson ration nu.

Missing docstring.

Missing docstring for get_effective_viscosity. Check Documenter's build log for details.

FastIsostasy.get_differential_fourier — Function

get_differential_fourier(W, N2)

Compute the matrices representing the differential operators in the fourier space.

Missing docstring.

Missing docstring for get_viscosity_ratio. Check Documenter's build log for details.

Missing docstring.

Missing docstring for three_layer_scaling. Check Documenter's build log for details.

FastIsostasy.loginterp_viscosity — Function

loginterp_viscosity(tvec, layer_viscosities, layers_thickness, pseudodiff)

Compute a log-interpolator of the equivalent viscosity from provided viscosity fields layer_viscosities at time stamps tvec.

Missing docstring.

Missing docstring for hyperbolic_channel_coeffs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_greenintegrand_coeffs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for build_greenintegrand. Check Documenter's build log for details.

FastIsostasy.get_elasticgreen — Function

get_elasticgreen(Omega, quad_support, quad_coeffs)

Integrate load response over field by using 2D quadrature with specified support points and associated coefficients.

FastIsostasy.get_quad_coeffs — Function

get_quad_coeffs(T, n)

Return support points and associated coefficients with specified Type for Gauss-Legendre quadrature.

FastIsostasy.quadrature1D — Function

quadrature1D(f, n, x1, x2)

Compute 1D Gauss-Legendre quadrature of f between x1 and x2 based on n support points.

Missing docstring.

Missing docstring for quadrature2D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_normalized_lin_transform. Check Documenter's build log for details.

Missing docstring.

Missing docstring for normalized_lin_transform. Check Documenter's build log for details.

FastIsostasy.kernelpromote — Function

kernelpromote(X, arraykernel)

Promote X to the kernel (Array or CuArray) specified by arraykernel.